CNV 19687: SidebarEditor for wizard overview & envs tabs
