--- src/utils/styles/_secondary-editor.scss
@@ -1,37 +1,0 @@
-.yaml-editor__buttons {
-  #save-changes {
-    width: 90px;
-    --pf-c-button--after--BorderColor: var(
-      --pf-c-button--m-secondary--after--BorderColor
-    ) !important;
-    color: var(--pf-c-button--m-secondary--Color) !important;
-    background-color: var(--pf-c-button--m-secondary--BackgroundColor) !important;
-  }
-
-  #reload-object {
-    --pf-c-button--after--BorderColor: var(
-      --pf-c-button--m-link--disabled--BackgroundColor
-    ) !important;
-    --pf-c-button--disabled--BackgroundColor: var(
-      --pf-c-button--m-link--disabled--BackgroundColor
-    ) !important;
-    color: var(--pf-c-button--m-link--Color) !important;
-    background-color: var(--pf-c-button--m-link--BackgroundColor) !important;
-  }
-
-  #cancel {
-    display: none;
-  }
-
-  // download button
-  .pf-c-button--align-right {
-    --pf-c-button--after--BorderColor: var(
-      --pf-c-button--m-link--disabled--BackgroundColor
-    ) !important;
-    --pf-c-button--disabled--BackgroundColor: var(
-      --pf-c-button--m-link--disabled--BackgroundColor
-    ) !important;
-    color: var(--pf-c-button--m-link--Color) !important;
-    background-color: var(--pf-c-button--m-link--BackgroundColor) !important;
-  }
-}
