+++ src/views/catalog/wizard/Wizard.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { useParams } from 'react-router-dom';
 
+import { SidebarEditorProvider } from '@kubevirt-utils/components/SidebarEditor/SidebarEditorContext';
 import { HorizontalNav } from '@openshift-console/dynamic-plugin-sdk';
 import { Stack, StackItem } from '@patternfly/react-core';
 
@@ -20,13 +21,15 @@ const Wizard: React.FC = () => {
   if (!vm) return <WizardEmptyState namespace={ns} />;
 
   return (
-    <Stack hasGutter>
-      <WizardHeader namespace={ns} />
-      <StackItem className="vm-wizard-body" isFilled>
-        <HorizontalNav pages={wizardNavPages} />
-      </StackItem>
-      <WizardFooter namespace={ns} />
-    </Stack>
+    <SidebarEditorProvider>
+      <Stack hasGutter>
+        <WizardHeader namespace={ns} />
+        <StackItem className="vm-wizard-body" isFilled>
+          <HorizontalNav pages={wizardNavPages} />
+        </StackItem>
+        <WizardFooter namespace={ns} />
+      </Stack>
+    </SidebarEditorProvider>
   );
 };
 export default Wizard;
