+++ src/views/catalog/wizard/tabs/environment/WizardEnvironmentTab.tsx
@@ -3,7 +3,9 @@ import * as React from 'react';
 import { WizardTab } from '@catalog/wizard/tabs';
 import EnvironmentForm from '@kubevirt-utils/components/EnvironmentEditor/EnvironmentForm';
 import Loading from '@kubevirt-utils/components/Loading/Loading';
-import { Bullseye } from '@patternfly/react-core';
+import { Bullseye, PageSection } from '@patternfly/react-core';
+
+import './wizard-environment-tab.scss';
 
 const WizardEnvironmentTab: WizardTab = ({ vm, updateVM, setDisableVmCreate }) => {
   if (!vm)
@@ -14,9 +16,9 @@ const WizardEnvironmentTab: WizardTab = ({ vm, updateVM, setDisableVmCreate }) =
     );
 
   return (
-    <div className="co-m-pane__body">
+    <PageSection className="wizard-environment-tab">
       <EnvironmentForm vm={vm} onEditChange={setDisableVmCreate} updateVM={updateVM} />
-    </div>
+    </PageSection>
   );
 };
