+++ src/views/catalog/wizard/tabs/environment/wizard-environment-tab.scss
@@ -1,0 +1,7 @@
+.wizard-environment-tab {
+    height: 100%;
+
+    .sidebar-editor .pf-c-sidebar__panel {
+        margin-bottom: var(--pf-global--spacer--md);
+    }
+}
