+++ src/views/catalog/wizard/tabs/overview/WizardOverviewTab.scss
@@ -1,3 +1,9 @@
 .wizard-overview-description-left-column {
   grid-column: 1;
 }
+
+.wizard-overview-tab {
+  &__grid {
+    margin-top: var(--pf-global--spacer--md);
+  }
+}
