+++ src/views/templates/details/tabs/details/TemplateDetailsPage.scss
@@ -9,7 +9,7 @@
 }
 
 .margin-top {
-  margin-top: 1rem;
+  margin-top: var(--pf-global--spacer--md);
 }
 
 .title-edit-item-space {
